collect-packages does not need a separate test file
