+++ src/Command.js
@@ -11,7 +11,7 @@ const Repository = require("./Repository");
 const writeLogFile = require("./utils/writeLogFile");
 const UpdatedPackagesCollector = require("./UpdatedPackagesCollector");
 const VersionSerializer = require("./VersionSerializer");
-const collectPackages = require("./utils/collectPackages");
+const collectPackages = require("./utils/collect-packages");
 const createPackageGraph = require("./utils/createPackageGraph");
 const filterPackages = require("./utils/filterPackages");
 const ValidationError = require("./utils/ValidationError");
