+++ test/fixtures/PackageUtilities/explicit-node-modules/package.json
@@ -1,5 +1,6 @@
 {
-  "name": "alle-pattern-monorepo",
+  "name": "alle-pattern-root",
+  "version": "1.0.0",
   "dependencies": {
     "package-3": "^1.0.0",
     "package-4": "^1.0.0"
