+++ test/fixtures/PackageUtilities/globstar/package.json
@@ -1,5 +1,6 @@
 {
-  "name": "globstar-monorepo",
+  "name": "globstar",
+  "version": "1.0.0",
   "dependencies": {
     "package-3": "^1.0.0",
     "package-4": "^1.0.0"
